+++ src/apps/seelen_rofi/index.tsx
@@ -2,10 +2,11 @@ import { getRootContainer } from '../shared';
 import { wrapConsole } from '../shared/ConsoleWrapper';
 import { App } from './App';
 import { registerDocumentEvents } from './events';
-import { initStore, store } from './store';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { declareDocumentAsLayeredHitbox } from 'seelen-core';
+
+import { initStore, store } from './modules/shared/store/infra';
 
 import './styles/reset.css';
 import './styles/colors.css';